server/traditionalTable: refactor mappings under fra/
