+++ server/traditionalTable/mappings/fra/annualReforestation.js
