+++ server/traditionalTable/mappings/fra/areaAffectedByFire.js
