+++ server/traditionalTable/mappings/fra/areaAffectedByFirePrint1.js
