+++ server/traditionalTable/mappings/fra/areaAffectedByFirePrint2.js
