+++ server/traditionalTable/mappings/fra/biomassStock.js
