+++ server/traditionalTable/mappings/fra/carbonStock.js
