+++ server/traditionalTable/mappings/fra/carbonStockSoilDepth.js
