+++ server/traditionalTable/mappings/fra/climaticDomain.js
