+++ server/traditionalTable/mappings/fra/degradedForest.js
