+++ server/traditionalTable/mappings/fra/disturbances.js
