+++ server/traditionalTable/mappings/fra/disturbancesPrint1.js
