+++ server/traditionalTable/mappings/fra/disturbancesPrint2.js
