+++ server/traditionalTable/mappings/fra/forestAreaWithinProtectedAreas.js
