+++ server/traditionalTable/mappings/fra/forestPolicy.js
