+++ server/traditionalTable/mappings/fra/graduationOfStudents.js
