+++ server/traditionalTable/mappings/fra/graduationOfStudentsPrint1.js
