+++ server/traditionalTable/mappings/fra/graduationOfStudentsPrint2.js
