+++ server/traditionalTable/mappings/fra/growingStockComposition.js
