+++ server/traditionalTable/mappings/fra/holderOfManagementRights.js
