+++ server/traditionalTable/mappings/fra/nonWoodForestProductsRemovalsCurrency.js
