+++ server/traditionalTable/mappings/fra/otherLandWithTreeCover.js
