+++ server/traditionalTable/mappings/fra/specificForestCategories.js
