+++ server/traditionalTable/mappings/fra/totalAreaWithDesignatedManagementObjective.js
