+++ server/traditionalTable/tableMappings.js
@@ -1,41 +1,41 @@
 const R = require('ramda')
 const assert = require('assert')
-const specificForestCategories = require('./mappings/specificForestCategories')
-const forestAreaChange = require('./mappings/forestAreaChange')
-const holderOfManagementRights = require('./mappings/holderOfManagementRights')
-const primaryDesignatedManagementObjective = require('./mappings/primaryDesignatedManagementObjective')
-const areaAffectedByFire = require('./mappings/areaAffectedByFire')
-const areaAffectedByFirePrint1 = require('./mappings/areaAffectedByFirePrint1')
-const areaAffectedByFirePrint2 = require('./mappings/areaAffectedByFirePrint2')
-const growingStockComposition = require('./mappings/growingStockComposition')
-const nonWoodForestProductsRemovals = require('./mappings/nonWoodForestProductsRemovals')
-const nonWoodForestProductsRemovalsCurrency = require('./mappings/nonWoodForestProductsRemovalsCurrency')
-const biomassStock = require('./mappings/biomassStock')
-const carbonStock = require('./mappings/carbonStock')
-const carbonStockSoilDepth = require('./mappings/carbonStockSoilDepth')
-const degradedForest = require('./mappings/degradedForest')
-const employment = require('./mappings/employment')
-const employmentPrint1 = require('./mappings/employmentPrint1')
-const employmentPrint2 = require('./mappings/employmentPrint2')
-const graduationOfStudents = require('./mappings/graduationOfStudents')
-const graduationOfStudentsPrint1 = require('./mappings/graduationOfStudentsPrint1')
-const graduationOfStudentsPrint2 = require('./mappings/graduationOfStudentsPrint2')
-const forestAreaWithinProtectedAreas = require('./mappings/forestAreaWithinProtectedAreas')
-const totalAreaWithDesignatedManagementObjective = require('./mappings/totalAreaWithDesignatedManagementObjective')
-const annualReforestation = require('./mappings/annualReforestation')
-const forestOwnership = require('./mappings/forestOwnership')
-const disturbances = require('./mappings/disturbances')
-const disturbancesPrint1 = require('./mappings/disturbancesPrint1')
-const disturbancesPrint2 = require('./mappings/disturbancesPrint2')
-const areaOfPermanentForestEstate = require('./mappings/areaOfPermanentForestEstate')
-const forestPolicy = require('./mappings/forestPolicy')
-const otherLandWithTreeCover = require('./mappings/otherLandWithTreeCover')
-const climaticDomain = require('./mappings/climaticDomain')
-const sustainableDevelopmentAgencyIndicator = require('./mappings/sustainableDevelopmentAgencyIndicator')
-const sustainableDevelopmentAgencySubIndicator1 = require('./mappings/sustainableDevelopmentAgencySubIndicator1')
-const sustainableDevelopmentAgencySubIndicator2 = require('./mappings/sustainableDevelopmentAgencySubIndicator2')
-const sustainableDevelopmentAgencySubIndicator3 = require('./mappings/sustainableDevelopmentAgencySubIndicator3')
-const sustainableDevelopmentAgencySubIndicator4 = require('./mappings/sustainableDevelopmentAgencySubIndicator4')
+const specificForestCategories = require('./mappings/fra/specificForestCategories')
+const forestAreaChange = require('./mappings/fra/forestAreaChange')
+const holderOfManagementRights = require('./mappings/fra/holderOfManagementRights')
+const primaryDesignatedManagementObjective = require('./mappings/fra/primaryDesignatedManagementObjective')
+const areaAffectedByFire = require('./mappings/fra/areaAffectedByFire')
+const areaAffectedByFirePrint1 = require('./mappings/fra/areaAffectedByFirePrint1')
+const areaAffectedByFirePrint2 = require('./mappings/fra/areaAffectedByFirePrint2')
+const growingStockComposition = require('./mappings/fra/growingStockComposition')
+const nonWoodForestProductsRemovals = require('./mappings/fra/nonWoodForestProductsRemovals')
+const nonWoodForestProductsRemovalsCurrency = require('./mappings/fra/nonWoodForestProductsRemovalsCurrency')
+const biomassStock = require('./mappings/fra/biomassStock')
+const carbonStock = require('./mappings/fra/carbonStock')
+const carbonStockSoilDepth = require('./mappings/fra/carbonStockSoilDepth')
+const degradedForest = require('./mappings/fra/degradedForest')
+const employment = require('./mappings/fra/employment')
+const employmentPrint1 = require('./mappings/fra/employmentPrint1')
+const employmentPrint2 = require('./mappings/fra/employmentPrint2')
+const graduationOfStudents = require('./mappings/fra/graduationOfStudents')
+const graduationOfStudentsPrint1 = require('./mappings/fra/graduationOfStudentsPrint1')
+const graduationOfStudentsPrint2 = require('./mappings/fra/graduationOfStudentsPrint2')
+const forestAreaWithinProtectedAreas = require('./mappings/fra/forestAreaWithinProtectedAreas')
+const totalAreaWithDesignatedManagementObjective = require('./mappings/fra/totalAreaWithDesignatedManagementObjective')
+const annualReforestation = require('./mappings/fra/annualReforestation')
+const forestOwnership = require('./mappings/fra/forestOwnership')
+const disturbances = require('./mappings/fra/disturbances')
+const disturbancesPrint1 = require('./mappings/fra/disturbancesPrint1')
+const disturbancesPrint2 = require('./mappings/fra/disturbancesPrint2')
+const areaOfPermanentForestEstate = require('./mappings/fra/areaOfPermanentForestEstate')
+const forestPolicy = require('./mappings/fra/forestPolicy')
+const otherLandWithTreeCover = require('./mappings/fra/otherLandWithTreeCover')
+const climaticDomain = require('./mappings/fra/climaticDomain')
+const sustainableDevelopmentAgencyIndicator = require('./mappings/fra/sustainableDevelopmentAgencyIndicator')
+const sustainableDevelopmentAgencySubIndicator1 = require('./mappings/fra/sustainableDevelopmentAgencySubIndicator1')
+const sustainableDevelopmentAgencySubIndicator2 = require('./mappings/fra/sustainableDevelopmentAgencySubIndicator2')
+const sustainableDevelopmentAgencySubIndicator3 = require('./mappings/fra/sustainableDevelopmentAgencySubIndicator3')
+const sustainableDevelopmentAgencySubIndicator4 = require('./mappings/fra/sustainableDevelopmentAgencySubIndicator4')
 
 const panEuropean = require('./mappings/panEuropean')
